changed comparator criteria, test failing
